+++ test/unit/logger.spec.js
@@ -1,11 +1,12 @@
 const { wrap } = require("../../src/logger");
 
 function callLogMethods(logger) {
-	logger.log("log level");
-	logger.info("info level");
+	logger.trace("trace level");
 	logger.debug("debug level");
+	logger.info("info level");
 	logger.warn("warn level");
 	logger.error("error level");	
+	logger.fatal("fatal level");
 }
 
 // Unit: OK!
@@ -18,54 +19,91 @@ describe("Test wrap", () => {
 			warn: jest.fn()
 		};
 		
-		let logger = wrap(con, null, "debug");
-		expect(logger.log).toBeInstanceOf(Function);
+		let logger = wrap(con, null, "trace");
+		expect(logger.fatal).toBeInstanceOf(Function);
 		expect(logger.error).toBeInstanceOf(Function);
 		expect(logger.warn).toBeInstanceOf(Function);
 		expect(logger.info).toBeInstanceOf(Function);
 		expect(logger.debug).toBeInstanceOf(Function);
+		expect(logger.trace).toBeInstanceOf(Function);
 
 		callLogMethods(logger);
 		expect(con.warn).toHaveBeenCalledTimes(1);
 		expect(con.warn).toHaveBeenCalledWith("warn level");
 
-		expect(con.error).toHaveBeenCalledTimes(1);
+		expect(con.error).toHaveBeenCalledTimes(2);
 		expect(con.error).toHaveBeenCalledWith("error level");
+		expect(con.error).toHaveBeenCalledWith("fatal level");
 
 		expect(con.info).toHaveBeenCalledTimes(3);
 		expect(con.info).toHaveBeenCalledWith("debug level");
 		expect(con.info).toHaveBeenCalledWith("info level");
-		expect(con.info).toHaveBeenCalledWith("log level");
+		expect(con.info).toHaveBeenCalledWith("trace level");
 	});
 
-	it("should create a full logger with moduleName", () => {
+	it("should create a full logger with missing methods", () => {
 		let con = {
-			debug: jest.fn(),
 			info: jest.fn(),
 			error: jest.fn(),
-			warn: jest.fn()
 		};
-		let logger = wrap(con, "Module", "debug");
-		expect(logger.log).toBeInstanceOf(Function);
+		
+		let logger = wrap(con, null, "trace");
+		expect(logger.fatal).toBeInstanceOf(Function);
 		expect(logger.error).toBeInstanceOf(Function);
 		expect(logger.warn).toBeInstanceOf(Function);
 		expect(logger.info).toBeInstanceOf(Function);
 		expect(logger.debug).toBeInstanceOf(Function);
+		expect(logger.trace).toBeInstanceOf(Function);
 
 		callLogMethods(logger);
 
-		expect(con.warn).toHaveBeenCalledTimes(1);
-		expect(con.warn).toHaveBeenCalledWith("[Module] warn level");
+		expect(con.error).toHaveBeenCalledTimes(3);
+		expect(con.error).toHaveBeenCalledWith("error level");
+		expect(con.error).toHaveBeenCalledWith("warn level");
+		expect(con.error).toHaveBeenCalledWith("fatal level");
+
+		expect(con.info).toHaveBeenCalledTimes(3);
+		expect(con.info).toHaveBeenCalledWith("debug level");
+		expect(con.info).toHaveBeenCalledWith("info level");
+		expect(con.info).toHaveBeenCalledWith("trace level");
+	});	
+
+	it("should create a full logger with moduleName", () => {
+		let con = {
+			trace: jest.fn(),
+			debug: jest.fn(),
+			info: jest.fn(),
+			warn: jest.fn(),
+			error: jest.fn(),
+			fatal: jest.fn()
+		};
+		let logger = wrap(con, "Module", "trace");
+		expect(logger.fatal).toBeInstanceOf(Function);
+		expect(logger.error).toBeInstanceOf(Function);
+		expect(logger.warn).toBeInstanceOf(Function);
+		expect(logger.info).toBeInstanceOf(Function);
+		expect(logger.debug).toBeInstanceOf(Function);
+		expect(logger.trace).toBeInstanceOf(Function);
+
+		callLogMethods(logger);
+
+		expect(con.fatal).toHaveBeenCalledTimes(1);
+		expect(con.fatal).toHaveBeenCalledWith("[Module] fatal level");
 
 		expect(con.error).toHaveBeenCalledTimes(1);
 		expect(con.error).toHaveBeenCalledWith("[Module] error level");
 
-		expect(con.info).toHaveBeenCalledTimes(2);
+		expect(con.warn).toHaveBeenCalledTimes(1);
+		expect(con.warn).toHaveBeenCalledWith("[Module] warn level");
+
+		expect(con.info).toHaveBeenCalledTimes(1);
 		expect(con.info).toHaveBeenCalledWith("[Module] info level");
-		expect(con.info).toHaveBeenCalledWith("[Module] log level");
 
 		expect(con.debug).toHaveBeenCalledTimes(1);
 		expect(con.debug).toHaveBeenCalledWith("[Module] debug level");
+
+		expect(con.trace).toHaveBeenCalledTimes(1);
+		expect(con.trace).toHaveBeenCalledWith("[Module] trace level");
 	});
 });
 
@@ -73,19 +111,23 @@ describe("Test wrap with logLevels", () => {
 
 	it("should create a filtered-level logger (warn)", () => {
 		let con = {
+			trace: jest.fn(),
 			debug: jest.fn(),
 			info: jest.fn(),
+			warn: jest.fn(),
 			error: jest.fn(),
-			warn: jest.fn()
+			fatal: jest.fn()
 		};
 		let logger = wrap(con, "Module", "warn");
-		expect(logger.log).toBeInstanceOf(Function);
+		expect(logger.fatal).toBeInstanceOf(Function);
 		expect(logger.error).toBeInstanceOf(Function);
 		expect(logger.warn).toBeInstanceOf(Function);
 		expect(logger.info).toBeInstanceOf(Function);
 		expect(logger.debug).toBeInstanceOf(Function);
+		expect(logger.trace).toBeInstanceOf(Function);
 
 		callLogMethods(logger);
+		expect(con.trace).toHaveBeenCalledTimes(0);
 		expect(con.debug).toHaveBeenCalledTimes(0);
 		expect(con.info).toHaveBeenCalledTimes(0);
 
@@ -93,14 +135,18 @@ describe("Test wrap with logLevels", () => {
 		expect(con.warn).toHaveBeenCalledWith("[Module] warn level");
 		expect(con.error).toHaveBeenCalledTimes(1);
 		expect(con.error).toHaveBeenCalledWith("[Module] error level");
+		expect(con.fatal).toHaveBeenCalledTimes(1);
+		expect(con.fatal).toHaveBeenCalledWith("[Module] fatal level");
 	});	
 
 	it("should create a filtered-level logger (module error)", () => {
 		let con = {
+			trace: jest.fn(),
 			debug: jest.fn(),
 			info: jest.fn(),
+			warn: jest.fn(),
 			error: jest.fn(),
-			warn: jest.fn()
+			fatal: jest.fn(),
 		};
 		let logger = wrap(con, "CTX", {
 			"*": "debug",
@@ -112,14 +158,17 @@ describe("Test wrap with logLevels", () => {
 		expect(con.info).toHaveBeenCalledTimes(0);
 		expect(con.warn).toHaveBeenCalledTimes(0);
 		expect(con.error).toHaveBeenCalledTimes(1);
+		expect(con.fatal).toHaveBeenCalledTimes(1);
 	});		
 
 	it("should create a filtered-level logger ('*' info)", () => {
 		let con = {
+			trace: jest.fn(),
 			debug: jest.fn(),
 			info: jest.fn(),
+			warn: jest.fn(),
 			error: jest.fn(),
-			warn: jest.fn()
+			fatal: jest.fn(),
 		};
 		let logger = wrap(con, "SVC", {
 			"*": "info",
@@ -127,18 +176,22 @@ describe("Test wrap with logLevels", () => {
 		});
 
 		callLogMethods(logger);
+		expect(con.trace).toHaveBeenCalledTimes(0);
 		expect(con.debug).toHaveBeenCalledTimes(0);
-		expect(con.info).toHaveBeenCalledTimes(2);
+		expect(con.info).toHaveBeenCalledTimes(1);
 		expect(con.warn).toHaveBeenCalledTimes(1);
 		expect(con.error).toHaveBeenCalledTimes(1);
+		expect(con.fatal).toHaveBeenCalledTimes(1);
 	});		
 
 	it("should create an empty logger (false)", () => {
 		let con = {
+			trace: jest.fn(),
 			debug: jest.fn(),
 			info: jest.fn(),
+			warn: jest.fn(),
 			error: jest.fn(),
-			warn: jest.fn()
+			fatal: jest.fn()
 		};
 		let logger = wrap(con, "SVC", {
 			"*": "info",
@@ -146,10 +199,12 @@ describe("Test wrap with logLevels", () => {
 		});
 
 		callLogMethods(logger);
+		expect(con.trace).toHaveBeenCalledTimes(0);
 		expect(con.debug).toHaveBeenCalledTimes(0);
 		expect(con.info).toHaveBeenCalledTimes(0);
 		expect(con.warn).toHaveBeenCalledTimes(0);
 		expect(con.error).toHaveBeenCalledTimes(0);
+		expect(con.fatal).toHaveBeenCalledTimes(0);
 	});		
 
 });